// src/user/interfaces/user.interface.ts
export interface UserInterfaces {

    id: number;
    code: string;
    firstName: string;
    lastName: string;
    username: string;
    password: string;
    email: string;
    phoneNo: string;
    mobileNo: string;
    active: boolean;
    createdBy: number;
    createdAt: Date;
    updatedBy: number;
    updatedAt: Date;
    deletedBy: number;
    deletedAt: Date;

}